Fix Jenkins RBAC to use ClusterRole
Jenkins runs in Jenkins namespace. We do not run apps in this namespace.
In order for jenkins to be able to use kubectl to run and stop pods in
other namespaces, where the apps run, RBAC needs a ClusterRole here
instead of a Role (which can only do actions in the jenkins namespace).