Consolidate service definitions
We need the service definition set as jenkins so that the jnlp slaves
can connect correctly. The reason for this is that they assume the
service for 8080 and 50000 is on the same domain.

We also only need internal service IP as jenkins will be exposed via a
proxy or LB.