add jenkins-secret-env-vars for global secrets
We have a requirement to add secrets for jobs to use. For instance,
service users for 3rd party services. Jenkins doesni't have a good way of
handling this so we mount them into the container at startup and set
them as variables. This probably needs a rethink as anyone with admin
access can see these secrets.